Creating new RemoteResourceHandler
