+++ TDS.ItemRenderer/src/main/java/tds/itemrenderer/data/ITSContent.java
@@ -8,13 +8,18 @@
  ******************************************************************************/
 package tds.itemrenderer.data;
 
-import java.util.ArrayList;
-import java.util.List;
-
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import com.sun.org.apache.xerces.internal.dom.ElementImpl;
+import com.sun.org.apache.xerces.internal.dom.ElementNSImpl;
+import com.sun.org.apache.xerces.internal.dom.NodeImpl;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.collections.Predicate;
 import org.apache.commons.lang.StringUtils;
 import org.w3c.dom.Element;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import tds.itemrenderer.data.apip.APIPXml;
 
@@ -225,7 +230,6 @@ public class ITSContent
 
   private ITSAttachment findBrailleAttachment(List<String> brailleSubTypeCodes, boolean isFindTranscript) {
     ITSAttachment brailleAttachment = null;
-
     // need to check these in order in case the content has both Math specific and No Math Code Needed versions
     for (String subTypeCode : brailleSubTypeCodes) {
       final String subTypeCodeFinal = subTypeCode + ( isFindTranscript ? "_transcript" : "" );
