+++ pom.xml
@@ -9,7 +9,7 @@
   <parent>
     <groupId>org.opentestsystem.shared</groupId>
     <artifactId>shared-master</artifactId>
-    <version>3.1.2-SNAPSHOT</version>
+    <version>3.1.2</version>
   </parent>
 
   <modules>
